requirements.txt
openai>=1.0.0
streamlit>=1.28.0
pandas>=1.5.0
numpy>=1.24.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0
requests>=2.31.0
python-dotenv>=1.0.0
PyPDF2>=3.0.0
python-docx>=0.8.11 